requirements.txt
six>=1.6.0,<1.7.0
Django>=3.2.6,<3.3.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0<5.4.0

flake8>=3.6.0,<3.7.0
pytest-django>=4.0.0,<4.4.0
pytest-xdist>2.0.0,<2.4.0
gunicorn>=20.1.0,<20.2.0
whitenoise>=5.3.0,<5.4.0
